Refactor form routing logic and integrate form configuration
- Moved form routing configurations to a separate file (form-config.js) for better organization and maintainability.
- Updated form-router.js and ms-form-router.js to import routing logic from form-config.js, simplifying the codebase.
- Enhanced field tracking initialization and error handling during form submission.
- Improved redirection logic to ensure proper handling of form submission outcomes.
