+++ index-ms.html
@@ -1,6 +1,8 @@
 <script charset="utf-8" type="text/javascript" src="//js.hsforms.net/forms/embed/v2.js"></script>
 <!-- Sentry (copy from your Sentry project settings) -->
 <script src="https://js.sentry-cdn.com/your-key.min.js" crossorigin="anonymous"></script>
+<!-- Load form router config first -->
+<script src="src/form-config.js"></script>
 <!-- Replace with JS Deliver Version for Prod Env-->
 <script src="src/ms-form-router.js"></script>
 <script>
